api2/pyproject.toml
[tool.poetry]
name = "boppop_api"
version = "0.1.0"
description = "Bop Pop API"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
strawberry-graphql = {extras = ["fastapi"], version = "0.257.0"}
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
python-dotenv = "^1.0.0"
alembic = "^1.12.1"
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.0.3"
redis = {extras = ["hiredis"], version = "^5.0.1"}
requests = "^2.31.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^14.0.1"}
asyncpg = "^0.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
